+++ changelog.md
@@ -10,6 +10,7 @@
 - Allowed the pizza minigame to be initiated without having to find or visit any secret rooms
 - Fixed Entropynnium being completely nonfunctional against Lich's 2nd phase
 - Prevented mastered Alien Nailgun from auto-spawning a few enemies that do more harm than good (e.g., enemies that summon other enemies)
+- Changed Chest Scanner to work on the player's currently targeted chest, rather than a random chest in the player's current room
 
 ## 1.32.4 (2026-07-05)